requirements.txt
certifi==2017.11.5
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.3
django-ckeditor==5.3.0
django-cors-headers==2.1.0
django-filter==1.0.4
django-js-asset==0.1.1
djangorestframework==3.6.3
geopy==1.11.0
idna==2.6
Markdown==2.6.8
oauthlib==2.0.6
olefile==0.44
Pillow==4.2.1
PyJWT==1.5.3
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.0.0
social-auth-core==1.5.0
urllib3==1.22
xlrd==1.0.0
